feat(export): move common calculation out of go-button
